core/requirements.txt
lmdb
argon2-cffi
cryptography
pycryptodome
rocksdict